requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.26.0
requests==2.32.2
python-dotenv==1.0.1
pydantic==2.7.1
